add utils file per Akshay suggestions
